Made sure code wouldn't crash for Sandbox mode.  Modified Seeker to see map in Sandbox.
