--- app/src/main/java/com/cascadia/hidenseek/TempToHome.java
@@ -1,49 +1,0 @@
-package com.cascadia.hidenseek;
-
-import android.app.Activity;
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.Button;
-import android.widget.ImageButton;
-
-public class TempToHome extends Activity {
-private Button btnHost;
-	@Override
-	protected void onCreate(Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
-		setContentView(R.layout.endofgameactivity);
-		btnHost = (Button) findViewById(R.id.home_button);
-		btnHost.setOnClickListener(new View.OnClickListener() {
-			public void onClick(View v) {
-				Intent intent = new Intent(TempToHome.this, Home.class);
-				startActivity(intent);
-				LoginManager.resetLoginManger();
-			
-			}
-		});
-		
-	}
-
-	@Override
-	public boolean onCreateOptionsMenu(Menu menu) {
-		// Inflate the menu; this adds items to the action bar if it is present.
-		//getMenuInflater().inflate(R.menu.temp_to_home, menu);
-		return true;
-	}
-	
-
-	@Override
-	public boolean onOptionsItemSelected(MenuItem item) {
-		// Handle action bar item clicks here. The action bar will
-		// automatically handle clicks on the Home/Up button, so long
-		// as you specify a parent activity in AndroidManifest.xml.
-		int id = item.getItemId();
-		if (id == R.id.action_settings) {
-			return true;
-		}
-		return super.onOptionsItemSelected(item);
-	}
-}
