+++ app/src/main/java/com/cascadia/hidenseek/login/HelpDialog.java
@@ -1,4 +1,4 @@
-package com.cascadia.hidenseek;
+package com.cascadia.hidenseek.login;
 
 import java.lang.String;
