+++ app/src/main/java/com/cascadia/hidenseek/login/HostLogin.java
@@ -1,4 +1,4 @@
-package com.cascadia.hidenseek;
+package com.cascadia.hidenseek.login;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -12,7 +12,11 @@ import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.Spinner;
 
-import com.cascadia.hidenseek.Player.Role;
+import com.cascadia.hidenseek.ConnectionChecks;
+import com.cascadia.hidenseek.R;
+import com.cascadia.hidenseek.model.Match;
+import com.cascadia.hidenseek.model.Player;
+import com.cascadia.hidenseek.model.Player.Role;
 import com.cascadia.hidenseek.network.PostMatchRequest;
 import com.cascadia.hidenseek.network.PostPlayerRequest;
 import com.cascadia.hidenseek.network.PutRoleRequest;
