+++ app/src/main/java/com/cascadia/hidenseek/login/JoinLogin.java
@@ -1,4 +1,4 @@
-package com.cascadia.hidenseek;
+package com.cascadia.hidenseek.login;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -11,6 +11,10 @@ import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 
+import com.cascadia.hidenseek.ConnectionChecks;
+import com.cascadia.hidenseek.R;
+import com.cascadia.hidenseek.model.Match;
+import com.cascadia.hidenseek.model.Player;
 import com.cascadia.hidenseek.network.GetMatchRequest;
 import com.cascadia.hidenseek.network.PostPlayerRequest;
