+++ app/src/main/java/com/cascadia/hidenseek/model/Player.java
@@ -1,4 +1,4 @@
-package com.cascadia.hidenseek;
+package com.cascadia.hidenseek.model;
 
 import android.location.Location;
