+++ app/src/main/java/com/cascadia/hidenseek/network/PutStartRequest.java
@@ -1,6 +1,6 @@
 package com.cascadia.hidenseek.network;
 
-import com.cascadia.hidenseek.Match;
+import com.cascadia.hidenseek.model.Match;
 import com.cascadia.hidenseek.network.NetworkBase.RequestType;
 
 public abstract class PutStartRequest extends NetworkRequest {
