+++ app/src/main/res/layout/activity_active.xml
@@ -3,7 +3,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.cascadia.hidenseek.Active"
+    tools:context="com.cascadia.hidenseek.active.Active"
 
     tools:ignore="MergeRootFrame" >
