+++ app/src/main/res/layout/activity_game_end.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:gravity="center"
+    android:orientation="vertical"
+    android:background="@drawable/bg" >
+
+    <TextView
+        android:id="@+id/textView1"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:gravity="center"
+        style="@style/CustomCastTheme"
+        android:text="@string/endGameMessage"
+        android:textAppearance="@style/TextAppearance.AppCompat.Large"
+        android:textColor="@color/ActiveText" />
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        android:gravity="center"
+        android:paddingTop="@dimen/activity_vertical_margin">
+
+        <Button
+            android:id="@+id/home_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/goHome"
+            android:textAppearance="@android:style/TextAppearance.Medium.Inverse"
+            android:layout_margin="@dimen/activity_vertical_margin" />
+
+        <Button
+            android:id="@+id/feedback_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/leaveFeedback"
+            android:textAppearance="@android:style/TextAppearance.Medium.Inverse"
+            android:layout_margin="@dimen/activity_vertical_margin" />
+    </LinearLayout>
+
+</LinearLayout>
