+++ app/src/main/res/layout/activity_host_config.xml
@@ -3,7 +3,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.cascadia.hidenseek.HostConfig"
+    tools:context="com.cascadia.hidenseek.login.HostConfig"
     tools:ignore="MergeRootFrame" >
 
     <ImageView
