+++ app/src/main/res/layout/activity_join_login.xml
@@ -4,7 +4,7 @@
     android:id="@+id/join_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.cascadia.hidenseek.HostLogin"
+    tools:context="com.cascadia.hidenseek.login.HostLogin"
     tools:ignore="MergeRootFrame" >
 
     <ImageView
