+++ app/src/main/res/layout/activity_select_match.xml
@@ -3,7 +3,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.cascadia.hidenseek.SelectMatch"
+    tools:context="com.cascadia.hidenseek.login.SelectMatch"
     tools:ignore="MergeRootFrame" >
 
     <ImageView
