+++ app/src/main/res/layout/settings.xml
@@ -3,7 +3,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.cascadia.hidenseek.Settings"
+    tools:context="com.cascadia.hidenseek.pending.Settings"
     tools:ignore="MergeRootFrame" >
     
     <ImageView
