--- app/src/main/res/layout/splash_screen.xml
@@ -1,51 +1,0 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/container"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    
-    tools:context="com.cascadia.hidenseek.Home"
-    tools:ignore="MergeRootFrame" >
-
-    <ImageView
-        android:contentDescription="@string/descr"
-        android:id="@+id/imageView1"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:scaleType="fitXY"
-        android:src="@drawable/bg" />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical" >
-		<ImageView
-		    android:contentDescription="@string/descr"
-            android:id="@+id/imageView2"
-            android:layout_width="137dp"
-            android:layout_height="121dp"
-            android:layout_gravity="center_horizontal"
-            android:layout_marginTop="16dp"
-            android:src="@drawable/logo_home" />
-
-		<TextView
-		    android:id="@+id/mTextField"
-		    android:layout_width="fill_parent"
-		    android:layout_height="wrap_content"
-		    android:layout_gravity="center"
-		    android:gravity="center"
-		    android:text="@string/app_name"
-		    android:textAppearance="?android:attr/textAppearanceLarge"
-		    android:textColor="@color/Players"
-		    android:textSize="@dimen/Counter"
-		    android:textStyle="bold" />
-
-		<ProgressBar
-		    android:id="@+id/progressBar1"
-		    style="?android:attr/progressBarStyleLarge"
-		    android:layout_width="fill_parent"
-		    android:layout_height="wrap_content" />
-
-	</LinearLayout>
-</FrameLayout>
