+++ app/src/main/res/menu/host_config.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.cascadia.hidenseek.HostConfig" >
+    tools:context="com.cascadia.hidenseek.login.HostConfig" >
 
     <item
         android:id="@+id/action_settings"
