+++ app/src/main/res/menu/host_login.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.cascadia.hidenseek.HostLogin" >
+    tools:context="com.cascadia.hidenseek.login.HostLogin" >
 
     <item
         android:id="@+id/action_settings"
