+++ app/src/main/res/menu/join_login.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.cascadia.hidenseek.JoinLogin" >
+    tools:context="com.cascadia.hidenseek.login.JoinLogin" >
 
     <item
         android:id="@+id/action_settings"
