+++ app/src/main/res/menu/select_match.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.cascadia.hidenseek.SelectMatch" >
+    tools:context="com.cascadia.hidenseek.login.SelectMatch" >
 
     <item
         android:id="@+id/action_settings"
