+++ app/src/main/res/menu/temp_to_home.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.cascadia.hidenseek.TempToHome" >
+    tools:context="com.cascadia.hidenseek.GameEnd" >
 
     <item
         android:id="@+id/action_settings"
