+++ app/src/main/res/values/colors.xml
@@ -2,5 +2,5 @@
         <color name="headings">#FEB41C</color>
         <color name="text">#EFFF</color>
         <color name="Players">#FFF</color>
-        <color name="ActiveText">#FEB41E</color>
+        <color name="ActiveText">#feb41e</color>
     </resources>
